% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bambaf_from_vcf.R
\name{bambaf_from_vcf}
\alias{bambaf_from_vcf}
\title{Get BAM baf information from vcf}
\usage{
bambaf_from_vcf(dir = ".", vcf_list, chr = NULL, projectname = "")
}
\arguments{
\item{dir}{The directory to all the vcf stored; default is right in this folder.}

\item{vcf_list}{All the vcf names stored in vcf.list; could use command:"ls *.vcf > vcf.list" to generate.}

\item{chr}{Specify the chromosome you want to generate. Must be of int from 1-22. If not specify, this function will generate all chromosomes.}

\item{projectname}{Name of the project}
}
\value{
void
}
\description{
If your vcf follow the format in the example, you could use this function to extract NGS baf from vcf files. Remember to load library before hands.
Save 6 lists, each list has N entry. N = # of individuals (or vcf file)
ngs_baf.nm: name of the bamfiles; ngs_baf.chr: the chromosome; ngs_baf.pos: the position of the variants; 
ngs_baf: the BAF of the variants; ngs_baf.id: the ID of the variants; filenm:the file name
}
\examples{
load('icnv.demobambaf_22.rda')
str(ngs_baf)
str(ngs_baf.pos)
\dontrun{
dir='PATH/TO/FOLDER'
bambaf_from_vcf(dir,'example_vcf.list')
bambaf_from_vcf(dir,'example_vcf.list',chr=22,projectname='icnv.demo')
}
}
