% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_array_input.R
\name{get_array_intput}
\alias{get_array_intput}
\title{Get array information from given format}
\usage{
get_array_intput(dir, pattern, chr = NULL, projectname = "")
}
\arguments{
\item{dir}{A string. The directory path to the folder where store signal intensity file according to chr}

\item{pattern}{A string. The pattern of all the intensity file}

\item{chr}{Specify the chromosome you want to generate. Must be of int from 1-22. If not specify, this function will generate files for all chromosomes.}

\item{projectname}{Name of the project}
}
\value{
void
}
\description{
If your array input file follow the format in the example, you could use this function to extract array LRR and baf. Remember to load library before hands.
Save 4*[# of chr] lists, each list has N entry. N = # of individuals
snp_lrr: SNP LRR intensity; snp_lrr.pos: the position of the SNPs
snp_baf: the BAF of the SNPs; snp_baf.pos: the position of the SNPs
}
\examples{
load('icnv.demoarray_lrrbaf_22.rda')
str(snp_lrr)
str(snp_lrr.pos)
str(snp_baf)
str(snp_baf.pos)
\dontrun{
dir='PATH/TO/FOLDER'
pattern=paste0('*.csv.arrayicnv$')
icnv_array_intput(dir,pattern,chr=22,projectname='icnv.demo')
}
}
