% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCNV_detection_function.R
\name{iCNV_detection}
\alias{iCNV_detection}
\title{Copy number variation detection tool for germline data. Able to combine intensity and BAF from SNP array and NGS data.}
\usage{
iCNV_detection(r1L = NULL, r2L = NULL, baf1 = NULL, baf2 = NULL,
  rpos1 = NULL, rpos2 = NULL, bpos1 = NULL, bpos2 = NULL, maxIt = 50,
  visual = 0, projname = "iCNV", CN = 0, mu = c(-3, 0, 2),
  cap = FALSE)
}
\arguments{
\item{r1L}{A list of NGS intensity data. Each entry is an individual. If no NGS data, no need to specify.}

\item{r2L}{A list of SNP array intensity data. Each entry is an individual. If no SNP array data, no need to specify.}

\item{baf1}{A list of NGS BAF data. Each entry is an individual. If no NGS data, no need to specify.}

\item{baf2}{A list of SNP array BAF data. Each entry is an individual. If no SNP array data, no need to specify.}

\item{rpos1}{A list of NGS intensity postion data. Each entry is an individual with dimension= (#of bins or exons, 2(start and end position)). If no NGS data, no need to specify.}

\item{rpos2}{A list of SNP array intensity postion data. Each entry is an individual with length=#of SNPs. If no SNP array data, no need to specify.}

\item{bpos1}{A list of NGS BAF postion data. Each entry is an individual with length=#of BAFs. If no NGS data, no need to specify.}

\item{bpos2}{A list of SNP array BAF postion data. Each entry is an individual with length=#of BAFs. If no SNP array data, no need to specify.}

\item{maxIt}{An integer number indicate the maximum number of EM iteration if not converged during parameter inference. Default 50.}

\item{visual}{An indicator variable with value {0,1,2}. 0 indicates no visualization, 1 indicates complete visualization, 2 indicates basic visualization}

\item{projname}{A string as the name of this project. Default 'iCNV'}

\item{CN}{An indicator variable with value {0,1} for whether wants to infer exact copy number. 0 no exact CN, 1 exact CN. Default 0.}

\item{mu}{A length tree vectur specify means of intensity in mixture normal distribution (Deletion, Diploid, Duplification). Default c(-3,0,2)}
}
\value{
(1) CNV inference, contains CNV inference, Start and end position for each inference, Conditional probability for each inference, mu for mixture normal, sigma for mixture normal, probability of CNVs, Z score for each inference.

(2) exact copy number for each CNV inference, if CN=1.
}
\description{
Copy number variation detection tool for germline data. Able to combine intensity and BAF from SNP array and NGS data.
}
\examples{
iCNV_detection()
}
\keyword{BAF,}
\keyword{CNV,}
\keyword{Intensity}
\keyword{Platform}
\keyword{integration,}
