\docType{data}
\name{icnv_res0}
\alias{icnv_res0}
\title{
    Example iCNV calling results.
}
\format{A list containing the calling result of CNVs:
    \describe{
        \item{1st item}{HMM call result without Copy number}
        \item{2nd item}{exact copy number}
}}
\usage{
    icnv_res
}
\value{
    iCNV calling result
}
\description{
    iCNV calling result of all the samples
}
\keyword{datasets}