\docType{data}
\name{ngs_baf.chr}
\alias{ngs_baf.chr}
\title{
    BAF chromosome from NGS
}
\format{A list of 46, which each entry is the BAF chromosome for a individual position
}
\usage{
    ngs_baf.chr
}
\value{
    BAF chromosome
}
\description{
    46 samples BAF chromosome. Pre-computed using whole exome sequencing data of 46 HapMap samples.
}
\keyword{datasets}