\docType{data}
\name{ngs_baf.id}
\alias{ngs_baf.id}
\title{
    BAF variants id from NGS
}
\format{A list of 46, which each entry is the BAF variants id a individual position
}
\usage{
    ngs_baf.id
}
\value{
    BAF variants id
}
\description{
    46 samples BAF ids. Pre-computed using whole exome sequencing data of 46 HapMap samples.
}
\keyword{datasets}