\docType{data}
\name{ngs_baf.pos}
\alias{ngs_baf.pos}
\title{
    BAF position list from NGS
}
\format{A list of ten, which each entry is the BAF positions for a individual
}
\usage{
    ngs_baf.pos
}
\value{
    BAF position
}
\description{
    10 samples BAF position extracted from VCF files, value stored at ngs_baf
}
\keyword{datasets}