\docType{data}
\name{ngs_baf}
\alias{ngs_baf}
\title{
    BAF list from NGS
}
\format{A list of ten, which each entry is the BAF value for a individual
}
\usage{
    ngs_baf
}
\value{
    BAF value
}
\description{
    10 samples BAF value extracted from VCF files, location stored at ngs_baf.pos
}
\keyword{datasets}