\docType{data}
\name{ngs_plr.pos}
\alias{ngs_plr.pos}
\title{
    Exon location list from NGS
}
\format{A list of ten, which each entry is the Exon positions for a individual
}
\usage{
    ngs_plr.pos
}
\value{
    Exon position
}
\description{
    10 samples exon position extracted from BED files, value stored at ngs_plr
}
\keyword{datasets}