\docType{data}
\name{ngs_plr}
\alias{ngs_plr}
\title{
    Normalized Poisson likelihood ratio list from NGS
}
\format{A list of ten, which each entry is the PLR value for a individual, calculated from CODEX
}
\usage{
    ngs_plr
}
\value{
    PLR value
}
\description{
    10 samples PLR value from BAM calculated by CODEX, exon position stored at ngs_plr.pos
}
\keyword{datasets}