\name{normObj}
\alias{normObj}
\docType{data}
\title{
Demo data pre-stored for normObj.
}
\description{
Pre-stored normObj data for demonstration purposes.
}
\usage{normObj}
\value{
normObj demo data (list) pre-computed.
}
\details{
Pre-computed using whole exome sequencing data of 46 HapMap samples.
}
\author{
Zilu Zhou \email{zhouzilu@pennmedicine.upenn.edu}
}
\examples{
Yhat <- normObjDemo$Yhat
AIC <- normObjDemo$AIC
BIC <- normObjDemo$BIC
RSS <- normObjDemo$RSS
K <- normObjDemo$K
}
\keyword{datasets}