% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHMMscore1_function.R
\name{output_list}
\alias{output_list}
\title{Generate ouput list.}
\usage{
output = output_list(icnv_res,sampleid, CN=0)
#' @export
}
\arguments{
\item{testres}{CNV inference result. Output from iCNV_detection()}

\item{sampleid}{the name of the sample, same order as the input}

\item{CN}{An indicator variable with value {0,1} for whether exact copy number inferred in iCNV_detection. 0 no exact CN, 1 exact CN. Default 0.}

\value{
output CNV list of each individual
}
\description{
Generate ouput list. From iCNV results.
}
\examples{
output = output_list(icnv_res=icnv_res,sampleid=sample_id, CN=0)
}
