% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ngs_plr.R
\name{plot_ngs_plr}
\alias{plot_ngs_plr}
\title{plot out the NGS plr.}
\usage{
plot_ngs_plr(ngs_plr, chr)
}
\arguments{
\item{ngs_plr}{Specify the ngs_plr object generated by CODEX.}

\item{chr}{Specify the chromosome you want to generate. Must be of int from 1-22}
}
\value{
void
}
\description{
plot out the NGS plr.
}
\examples{
library(fields)
plot_ngs_plr(ngs_plr,chr)
}
