% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHMMscore_function.R
\name{plotHMMscore}
\alias{plotHMMscore}
\title{Plot CNV inference score.}
\usage{
plotHMMscore(icnv_res, h = min(icnv_res[[1]][[1]][[2]]),
  t = max(icnv_res[[1]][[1]][[2]]), subj = "score plot")
}
\arguments{
\item{icnv_res}{CNV inference result. Result from iCNV_detection() (i.e. iCNV_detection(...))}

\item{h}{start position of this plot. Default Start of the whole chromosome}

\item{t}{end position of this plot. Default End of the whole chromosome}

\item{title}{of this plot. Default "score plot"}
}
\value{
void
}
\description{
Plot CNV inference score.
}
\examples{
icnv_res = iCNV_detection(...)
pdf(file=paste0(projname,'.pdf'),width=13,height = 10)
plotHMMscore(icnv_res,h=100000, t=200000, subj='my favorite subject')
dev.off()
}
