% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHMMscore1_function.R
\name{plotHMMscore1}
\alias{plotHMMscore1}
\title{Plot CNV inference score.}
\usage{
plotHMMscore1(testres, h = min(testres[[1]][[2]]),
  t = max(testres[[1]][[2]]), subj = "score plot")
}
\arguments{
\item{testres}{CNV inference result. Output from iCNV_detection()}

\item{h}{start position of this plot. Default Start of the whole chromosome}

\item{t}{end position of this plot. Default End of the whole chromosome}

\item{title}{of this plot. Default "score plot"}
}
\value{
void
}
\description{
Plot CNV inference score.
}
\examples{
pdf(file=paste0(projname,'.pdf'),width=13,height = 10)
plotHMMscore1(testres,h=100000, t=200000, subj='my favorite subject')
dev.off()
}
