\name{qcObj}
\alias{qcObj}
\docType{data}
\title{
Demo data pre-stored for qcObj.
}
\description{
Pre-stored qcObj data for demonstration purposes.
}
\usage{qcObj}
\value{
qcObj demo data (list) pre-computed.
}
\details{
Pre-computed using whole exome sequencing data of 46 HapMap samples.
}
\author{
Zilu Zhou \email{zhouzilu@pennmedicine.upenn.edu}
}
\examples{
Y_qc <- qcObj$Y_qc
sampname_qc <- qcObj$sampname_qc
gc_qc <- qcObj$gc_qc
mapp_qc <- qcObj$mapp_qc
ref_qc <- qcObj$ref_qc
}
\keyword{datasets}