\docType{data}
\name{snp_baf.pos}
\alias{snp_baf.pos}
\title{
    BAF position list from Array
}
\format{A list of ten, which each entry is the BAF positions for a individual
}
\usage{
    snp_baf.pos
}
\value{
    BAF position
}
\description{
    10 samples BAF position extracted from standard format, value stored at snp_baf
}
\keyword{datasets}