\docType{data}
\name{snp_baf}
\alias{snp_baf}
\title{
    BAF list from Array
}
\format{A list of ten, which each entry is the BAF value for a individual
}
\usage{
    snp_baf
}
\value{
    BAF value
}
\description{
    10 samples BAF value extracted from standard format files, location stored at snp_baf.pos
}
\keyword{datasets}