\docType{data}
\name{snp_lrr.pos}
\alias{snp_lrr.pos}
\title{
    SNP position list from Array
}
\format{A list of ten, which each entry is the SNP positions for a individual
}
\usage{
    snp_lrr.pos
}
\value{
    SNP position
}
\description{
    10 samples SNP position extracted from standard format, value stored at snp_lrr
}
\keyword{datasets}