\docType{data}
\name{snp_lrr}
\alias{snp_lrr}
\title{
    Normalized log R ratio list from Array
}
\format{A list of ten, which each entry is the LRR value for a individual
}
\usage{
    snp_lrr
}
\value{
    LRR value
}
\description{
    10 samples LRR value from standard format, SNP position stored at snp_lrr.pos
}
\keyword{datasets}