% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_list_function.R
\name{output_list}
\alias{output_list}
\title{Generate ouput list.}
\usage{
output_list(icnv_res, sampleid = NULL, CN = 0, min_size = 0)
}
\arguments{
\item{icnv_res}{CNV inference result. Output from iCNV_detection()}

\item{sampleid}{the name of the sample, same order as the input}

\item{CN}{An indicator variable with value {0,1} for whether exact copy number inferred in iCNV_detection. 0 no exact CN, 1 exact CN. Default 0.}

\item{min_size}{A integer which indicate the minimum length of the CNV you are interested in. This could remove super short CNVs due to noise. Default 0. Recommend 1000.}
}
\value{
output CNV list of each individual
}
\description{
Generate human readable output from result calculated by iCNV_detection function
}
\examples{
icnv.output = output_list(icnv_res=icnv_res,sampleid=sampname_qc, CN=0)
}
