% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHMMscore_function.R
\name{plotHMMscore}
\alias{plotHMMscore}
\title{Plot CNV inference score.}
\usage{
plotHMMscore(HMMcalls, CNV, h = min(HMMcalls[[1]][[2]]),
  t = max(HMMcalls[[1]][[2]]), subj = "score plot")
}
\arguments{
\item{HMMcalls}{CNV inference result. First entry from iCNV_detection() result (i.e. iCNV_detection(...)[[1]])}

\item{CNV}{Genotype inference result. Second entry from iCNV_detection() result (i.e. iCNV_detection(...)[[2]])}

\item{h}{start position of this plot. Default Start of the whole chromosome}

\item{t}{end position of this plot. Default End of the whole chromosome}

\item{title}{of this plot. Default "score plot"}
}
\value{
void
}
\description{
Plot CNV inference score.
}
\examples{
testres = iCNV_detection(...)
HMMcalls = testres[[1]]
pdf(file=paste0(projname,'.pdf'),width=13,height = 10)
plotHMMscore1(HMMcalls,h=100000, t=200000, subj='my favorite subject')
dev.off()
}
